% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polish.R
\docType{data}
\name{Polish}
\alias{Polish}
\title{Capture-Recapture samples for people with a Polish nationality living in the
Netherlands in 2009.}
\format{
An object of class \code{data.frame} with 3 rows and 3 columns.
}
\usage{
Polish
}
\description{
Two registers of Statistics Netherlands, the GBA and the HKS. Van der Heijden,
P.G.M., M.J.L.F. Cruy, and G. van Gils. 2011. Aantallen Geregistreerde en
Nietgeregistreerde Burgers uit MOE-landen die in Nederland Verblijven.
Rapportage Schattingen 2008 en 2009. The Number of Registered and Non-registered
Citizens from MOE-countries Residing in the Netherlands. Reporting Estimations
2008 and 2009. The Hague: Ministry of Social Affairs and Employment. Available
at: http:// www.rijksoverheid.nl/documenten-en-publicaties/rapporten/2013/01/14/
aantallen-gere gistreerde-en-niet-geregistreerde-burgers-uit-moe-landen-die-
in-nederland-verblijven. html (in Dutch) (accessed July 2015).
}
\keyword{datasets}
