% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWID.R
\docType{data}
\name{PWID}
\alias{PWID}
\title{Capture-Recapture samples for People Who Inject Drugs in Brussels, Belgium.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 4 columns.
}
\usage{
PWID
}
\description{
Three capture-recapture samples of PWID in Brussels. See (Els Plettinckx, Forrest W. Crawford,
Jérôme Antoine, Gremeaux Lies, and Luk Van Baelen.
How many people injected drugs over the last 12 months in Belgium? Estimates based on a
capture-recapture and multiplier method. Working Paper, 2020.) and (Sun, Jinghao, Luk Van Baelen,
Els Plettinckx, and Forrest W. Crawford. "Partial identification and dependence-robust confidence
intervals for capture-recapture surveys." arXiv preprint arXiv:2008.00127 (2020))
(https://arxiv.org/abs/2008.00127) for details.
}
\keyword{datasets}
